bant/requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
msgpack==1.0.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
whitenoise==5.2.0
zope.interface==5.1.0
